refactor(today): extract snapshot builder
