+++ src/lib/features/today/snapshot.ts
@@ -1,14 +1,5 @@
 import type { HealthDbJournalEntriesStore } from '$lib/core/db/types';
-import type {
-  DailyRecord,
-  FoodCatalogItem,
-  FoodEntry,
-  HealthEvent,
-  JournalEntry,
-  PlanSlot,
-  PlannedMeal,
-  RecipeCatalogItem,
-} from '$lib/core/domain/types';
+import type { HealthEvent, JournalEntry } from '$lib/core/domain/types';
 import {
   buildDailyNutritionSummary,
   type NutritionRecommendationContextStore,
@@ -20,7 +11,6 @@ import {
 } from '$lib/features/nutrition/store';
 import {
   getNutritionPlannedMealResolution,
-  resolveNutritionPlannedMeal,
   type NutritionPlannedMealResolution,
   type NutritionPlannedMealStore,
 } from '$lib/features/nutrition/planned-meal-resolution';
@@ -29,15 +19,16 @@ import {
   listWorkoutTemplates,
   type MovementStorage,
 } from '$lib/features/movement/service';
-import { createSlotSummary } from '$lib/features/planning/model';
 import { listPlanSlotsForDay } from '$lib/features/planning/service';
-import {
-  buildTodayIntelligence,
-  buildTodayRecoveryAdaptation,
-  type TodayIntelligenceResult,
-  type TodayPlannedWorkoutInput as TodayPlannedWorkoutModel,
-  type TodayRecoveryAdaptationInput as TodayRecoveryAdaptationModel,
-} from '$lib/features/today/intelligence';
+import { buildTodaySnapshotFromData, type TodaySnapshot } from './snapshot-builder';
+
+export type {
+  TodayPlanItem,
+  TodayPlannedWorkout,
+  TodayRecoveryAdaptation,
+  TodaySnapshot,
+} from './snapshot-builder';
+export { buildTodaySnapshotFromData } from './snapshot-builder';
 
 export type JournalEntriesStore = HealthDbJournalEntriesStore;
 
@@ -48,39 +39,6 @@ export interface TodaySnapshotStore
     RecipeCatalogItemsStore,
     MovementStorage,
     JournalEntriesStore {}
-
-export interface TodayPlanItem {
-  id: string;
-  title: string;
-  subtitle: string;
-  status: PlanSlot['status'];
-}
-
-export type TodayPlannedWorkout = TodayPlannedWorkoutModel;
-
-export type TodayRecoveryAdaptation = TodayRecoveryAdaptationModel;
-
-export interface TodaySnapshot {
-  date: string;
-  dailyRecord: DailyRecord | null;
-  foodEntries: FoodEntry[];
-  nutritionSummary: {
-    calories: number;
-    protein: number;
-    fiber: number;
-    carbs: number;
-    fat: number;
-  };
-  plannedMeal: PlannedMeal | null;
-  plannedMealIssue: string | null;
-  plannedWorkout: TodayPlannedWorkout | null;
-  plannedWorkoutIssue: string | null;
-  recoveryAdaptation: TodayRecoveryAdaptation | null;
-  intelligence: TodayIntelligenceResult;
-  planItems: TodayPlanItem[];
-  events: HealthEvent[];
-  latestJournalEntry: JournalEntry | null;
-}
 
 export async function listEventsForDay(
   store: NutritionRecommendationContextStore,
@@ -98,176 +56,11 @@ async function getLatestJournalEntryForDay(
   );
 }
 
-function deriveTodayPlannedWorkout(
-  planSlots: PlanSlot[],
-  foodCatalogItems: FoodCatalogItem[],
-  recipeCatalogItems: Awaited<ReturnType<typeof listRecipeCatalogItems>>,
-  workoutTemplates: Awaited<ReturnType<typeof listWorkoutTemplates>>,
-  exerciseCatalogItems: Awaited<ReturnType<typeof listExerciseCatalogItems>>
-): TodayPlannedWorkout | null {
-  for (const slot of planSlots) {
-    if (slot.slotType !== 'workout' || slot.status !== 'planned') {
-      continue;
-    }
-
-    if (slot.itemType === 'workout-template' && slot.itemId) {
-      const template = workoutTemplates.find((candidate) => candidate.id === slot.itemId);
-      if (!template) {
-        continue;
-      }
-    }
-
-    return {
-      id: slot.id,
-      title: slot.title,
-      subtitle: createSlotSummary(
-        slot,
-        foodCatalogItems,
-        recipeCatalogItems,
-        workoutTemplates,
-        exerciseCatalogItems
-      ),
-      status: slot.status,
-    };
-  }
-
-  return null;
-}
-
-function deriveTodayPlannedWorkoutIssue(
-  planSlots: PlanSlot[],
-  workoutTemplates: Awaited<ReturnType<typeof listWorkoutTemplates>>
-): string | null {
-  for (const slot of planSlots) {
-    if (
-      slot.slotType !== 'workout' ||
-      slot.itemType !== 'workout-template' ||
-      !slot.itemId ||
-      slot.status !== 'planned'
-    ) {
-      continue;
-    }
-
-    const exists = workoutTemplates.some((template) => template.id === slot.itemId);
-    if (!exists) {
-      return 'That planned workout no longer exists. Replace it in Plan before using it today.';
-    }
-  }
-
-  return null;
-}
-
 export async function getTodayPlannedMealResolution(
   store: NutritionPlannedMealStore,
   date: string
 ): Promise<NutritionPlannedMealResolution> {
   return await getNutritionPlannedMealResolution(store, date);
-}
-
-export function buildTodaySnapshotFromData(input: {
-  date: string;
-  dailyRecord: DailyRecord | null;
-  nutritionSummary: {
-    calories: number;
-    protein: number;
-    fiber: number;
-    carbs: number;
-    fat: number;
-    entries: FoodEntry[];
-  };
-  planSlots: PlanSlot[];
-  foodCatalogItems: FoodCatalogItem[];
-  recipeCatalogItems: RecipeCatalogItem[];
-  workoutTemplates: import('$lib/core/domain/types').WorkoutTemplate[];
-  exerciseCatalogItems: import('$lib/core/domain/types').ExerciseCatalogItem[];
-  events: HealthEvent[];
-  latestJournalEntry: JournalEntry | null;
-}): TodaySnapshot {
-  const {
-    date,
-    dailyRecord,
-    nutritionSummary,
-    planSlots,
-    foodCatalogItems,
-    recipeCatalogItems,
-    workoutTemplates,
-    exerciseCatalogItems,
-    events,
-    latestJournalEntry,
-  } = input;
-
-  const planItems: TodayPlanItem[] = planSlots.map((slot) => ({
-    id: slot.id,
-    title: slot.title,
-    subtitle: createSlotSummary(
-      slot,
-      foodCatalogItems,
-      recipeCatalogItems,
-      workoutTemplates,
-      exerciseCatalogItems
-    ),
-    status: slot.status,
-  }));
-
-  const plannedWorkout = deriveTodayPlannedWorkout(
-    planSlots,
-    foodCatalogItems,
-    recipeCatalogItems,
-    workoutTemplates,
-    exerciseCatalogItems
-  );
-  const plannedMealResolution = resolveNutritionPlannedMeal(planSlots, foodCatalogItems);
-  const plannedWorkoutIssue = plannedWorkout
-    ? null
-    : deriveTodayPlannedWorkoutIssue(planSlots, workoutTemplates);
-  const recoveryAdaptation = buildTodayRecoveryAdaptation({
-    dailyRecord,
-    events,
-    plannedMeal: plannedMealResolution.candidate?.meal ?? null,
-    plannedWorkout,
-    foodCatalogItems,
-  });
-  const intelligence = buildTodayIntelligence({
-    date,
-    dailyRecord,
-    nutritionSummary: {
-      calories: nutritionSummary.calories,
-      protein: nutritionSummary.protein,
-      fiber: nutritionSummary.fiber,
-      carbs: nutritionSummary.carbs,
-      fat: nutritionSummary.fat,
-    },
-    plannedMeal: plannedMealResolution.candidate?.meal ?? null,
-    plannedMealIssue: plannedMealResolution.issue,
-    plannedWorkout,
-    plannedWorkoutIssue,
-    recoveryAdaptation,
-    latestJournalEntry,
-    events,
-    planItemsCount: planItems.length,
-  });
-
-  return {
-    date,
-    dailyRecord,
-    foodEntries: nutritionSummary.entries,
-    nutritionSummary: {
-      calories: nutritionSummary.calories,
-      protein: nutritionSummary.protein,
-      fiber: nutritionSummary.fiber,
-      carbs: nutritionSummary.carbs,
-      fat: nutritionSummary.fat,
-    },
-    plannedMeal: plannedMealResolution.candidate?.meal ?? null,
-    plannedMealIssue: plannedMealResolution.issue,
-    plannedWorkout,
-    plannedWorkoutIssue,
-    recoveryAdaptation,
-    intelligence,
-    planItems,
-    events,
-    latestJournalEntry,
-  };
 }
 
 async function buildTodaySnapshotData(
